[Cleanup] Merging PMachine with StateMachine
