+++ Tst/RegressionTests/Feature2Stmts/StaticError/newMachine1/newMachine1.p
@@ -1,0 +1,14 @@
+machine Main {
+  start state Init {
+    entry {
+      new M2();
+    }
+  }
+}
+
+machine M2 {
+  start state Init {
+    entry(payload : (n: int, s: string)) {
+    }
+  }
+}
