+++ Tst/RegressionTests/Feature2Stmts/StaticError/newMachine2/newMachine2.p
@@ -1,0 +1,14 @@
+machine Main {
+  start state Init {
+    entry {
+      new M2((n="payload string", s=1));
+    }
+  }
+}
+
+machine M2 {
+  start state Init {
+    entry { 
+    }
+  }
+}
