+++ Tst/RegressionTests/Feature2Stmts/StaticError/newMachine3/newMachine3.p
@@ -1,0 +1,15 @@
+
+machine Main {
+  start state Init {
+    entry {
+      new M2((n="payload string", s=1));
+    }
+  }
+}
+
+machine M2 {
+  start state Init {
+    entry(payload : (n: int, s: string)) {
+    }
+  }
+}
